Добавлен модуль компонента формы редактирования задачи
